+++ src/modules/notes/helper/index.ts
@@ -6,6 +6,7 @@ import * as pdf from 'pdf-parse'
 import { IConfigAttributes } from 'src/common/interfaces/config/app-config.interface'
 import { getConfig } from '../../../config'
 import { Note } from '../models/notes.model'
+import { Rating } from 'src/modules/ratings/models/rating.model'
 
 const config: IConfigAttributes = getConfig()
 
@@ -40,15 +41,6 @@ export async function calculateReadTimeMinutes(body: string): Promise<number> {
 }
 
 /**
- * Used to initialize a ratings format for a new note (stick to defaul)
- * @param ratingsSize The number of stars for this rating object
- * @returns An array of ratingSize initialized with zeroes
- */
-export function createEmptyRatings(ratingsSize = 5): number[] {
-	return Array(ratingsSize).fill(0)
-}
-
-/**
  * Overrides default file name and allows additional elements to be part of the file name
  * @param req The request object from the HTTP request
  * @param file The Multer File object containing a buffer and file metadata
@@ -58,6 +50,15 @@ export function editFileName(req, file: Express.Multer.File, cb) {
 	const name = file.originalname.split('.')[0]
 	const fileExtName = extname(file.originalname)
 	cb(null, `${name}-${uuid()}${fileExtName}`)
+}
+
+export function getRatingTotals(ratings: Rating[]): number[] {
+	const totals: number[] = [ 0, 0, 0, 0, 0 ]
+	for (const r of ratings) {
+		totals[r.value - 1] += 1
+	}
+
+	return totals
 }
 
 /**
@@ -77,8 +78,10 @@ export function compareNotesWithCombinedFeatures(a: Note, b: Note): number {
 	aCreated < bCreated ? (aPoints += 30) : (bPoints += 30)
 
 	// Now calculate points for each notes ratings
-	aPoints += a.rating.reduce((a, b, i) => a + b * i, 0)
-	bPoints += b.rating.reduce((a, b, i) => a + b * i, 0)
+	const aRatings: number[] = this.getRatingTotals(a.ratings)
+	const bRatings: number[] = this.getRatingTotals(b.ratings)
+	aPoints += aRatings.reduce((a, b, i) => a + b * i, 0)
+	bPoints += bRatings.reduce((a, b, i) => a + b * i, 0)
 
 	// Finally compare and return
 	return bPoints - aPoints